rename customInspect.ts
