+++ src/internal/app.ts
@@ -2,7 +2,7 @@ import type * as models from "../models/mod.ts";
 import type { BasicObjectInit, Creator } from "./client.ts";
 import { suppressAPIError } from "./api_error.ts";
 import { assert } from "../deps/std_testing_asserts.ts";
-import { customInspect } from "./customInspect.ts";
+import { customInspect } from "./custom_inspect.ts";
 
 export type AppData = models.APIData["getApps"][number];
