+++ src/internal/custom_inspect.ts
