+++ src/internal/location.ts
@@ -2,7 +2,7 @@ import { chunk } from "../deps/std_collections_chunk.ts";
 import type * as models from "../models/mod.ts";
 import type { BasicObjectInit, Creator } from "./client.ts";
 import { suppressAPIError } from "./api_error.ts";
-import { customInspect } from "./customInspect.ts";
+import { customInspect } from "./custom_inspect.ts";
 
 export type LocationData = models.APIData["getLocation"];
